[platform_device_loader] isolate update device related operation
