+++ libparsec/crates/platform_device_loader/src/device/mod.rs
@@ -4,8 +4,10 @@ mod archive;
 mod list;
 mod recovery;
 mod save;
+mod update;
 
 pub use archive::*;
 pub use list::*;
 pub use recovery::*;
 pub use save::*;
+pub use update::*;
